requirements.txt
# Pea Kingdom Bingo - Team Competition Platform
# Python dependencies for Flask web application

# Core Flask framework and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Caching==2.1.0

# Database support
SQLAlchemy==2.0.21

# AWS integration for authentication
boto3==1.28.85
botocore==1.31.85

# JWT token handling
PyJWT==2.8.0
cryptography==41.0.7

# HTTP requests for JWKS
requests==2.31.0

# WSGI server for production
gunicorn==21.2.0

# Development and testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
